Fix backwards test in mergeLists
* Add unit tests for mergeLists.
* Refactor mergeLists out of Source for ease of testing.
