+++ ui/src/data/source.tsx
@@ -85,11 +85,6 @@ export type FilterCountValue = {
   [x: string]: DataValue;
 };
 
-export type MergedItem<T> = {
-  source: string;
-  data: T;
-};
-
 export type PropertyMap = {
   [key: string]: string;
 };
@@ -247,13 +242,6 @@ export interface Source {
     filter: Filter | null,
     groupByAttributes?: string[]
   ): Promise<FilterCountValue[]>;
-
-  mergeLists<T, R>(
-    lists: [string, T[]][],
-    maxCount: number,
-    direction: SortDirection,
-    get: (value: T) => R
-  ): MergedItem<T>[];
 
   listCohortReviews(studyId: string, cohortId: string): Promise<CohortReview[]>;
 
@@ -682,48 +670,6 @@ export class BackendSource implements Source {
       processAttributes(value, count.attributes);
       return value;
     });
-  }
-
-  public mergeLists<T, R>(
-    lists: [string, T[]][],
-    maxCount: number,
-    direction: SortDirection,
-    get: (value: T) => R
-  ): MergedItem<T>[] {
-    const merged: MergedItem<T>[] = [];
-    const sources = lists.map(
-      ([source, data]) => new MergeSource<T>(source, data)
-    );
-    while (true) {
-      let maxSource: MergeSource<T> | undefined;
-
-      sources.forEach((source) => {
-        if (!source.done()) {
-          if (!maxSource) {
-            maxSource = source;
-          } else {
-            const value = get(source.peek());
-            const maxValue = get(maxSource.peek());
-            if (
-              (value && !maxValue) ||
-              (value &&
-                maxValue &&
-                (direction == SortDirection.Asc
-                  ? value > maxValue
-                  : value < maxValue))
-            ) {
-              maxSource = source;
-            }
-          }
-        }
-      });
-      if (!maxSource || merged.length === maxCount) {
-        break;
-      }
-      merged.push({ source: maxSource.source, data: maxSource.pop() });
-    }
-
-    return merged;
   }
 
   public async listCohortReviews(
@@ -1269,30 +1215,6 @@ export class BackendSource implements Source {
       limit: limit ?? 50,
     };
   }
-}
-
-class MergeSource<T> {
-  constructor(public source: string, private data: T[]) {
-    this.source = source;
-    this.data = data;
-    this.current = 0;
-  }
-
-  done() {
-    return this.current === this.data.length;
-  }
-
-  peek() {
-    return this.data[this.current];
-  }
-
-  pop(): T {
-    const data = this.peek();
-    this.current++;
-    return data;
-  }
-
-  private current: number;
 }
 
 function isInternalAttribute(attribute: string): boolean {
